backend/requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5
backcall==0.1.0
beautifulsoup4==4.6.0
branca==0.3.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==2.2.7
django-allauth==0.41.0
django-cors-headers==3.2.1
django-pandas==0.6.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-sslserver==0.22
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
folium==0.10.1
idna==2.8
inflection==0.3.1
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
itypes==1.1.0
jedi==0.16.0
Jinja2==2.10.3
joblib==0.14.0
JPype1==0.7.2
kiwisolver==1.1.0
konlpy==0.5.2
lazy-object-proxy==1.4.3
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
nltk==3.5
numpy==1.17.3
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.3
pandas==0.25.3
parso==0.6.2
pickleshare==0.7.5
Pillow==7.1.1
prompt-toolkit==3.0.5
pycodestyle==2.5.0
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
pyparsing==2.4.5
PySocks==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
regex==2020.4.4
requests==2.22.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scikit-learn==0.21.3
scikit-surprise==1.1.0
scipy==1.3.1
seaborn==0.9.0
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
tqdm==4.45.0
traitlets==4.3.3
tweepy==3.8.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.7
wcwidth==0.1.9
wordcloud==1.6.0
wrapt==1.11.2
